Implementation of bezier patche model generator (#50)

+++ src/lib/matrix.cpp
@@ -29,7 +29,7 @@ void cross(float *a, float *b, float *res) {
 
 
 void normalize(float *a) {
-        float l {sqrt(a[0] * a[0] + a[1] * a[1] + a[2] * a[2])};
+        float l {static_cast<float>(sqrt(a[0] * a[0] + a[1] * a[1] + a[2] * a[2]))};
         a[0] = a[0] / l;
         a[1] = a[1] / l;
         a[2] = a[2] / l;
@@ -37,7 +37,7 @@ void normalize(float *a) {
 
 
 float length(float *v) {
-        float res {sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2])};
+        float res {static_cast<float>(sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]))};
         return res;
 }
